% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{getDatasetFields}
\alias{getDatasetFields}
\title{Get Dataset Fields}
\usage{
getDatasetFields(dataset_id)
}
\arguments{
\item{dataset_id}{Dataset ID to query}
}
\value{
A \code{data.frame} containing the dataset metadata
}
\description{
Get the fields of a dataset by dataset ID
}
\seealso{
\url{https://developer.domo.com/docs/dataset-api-reference/dataset}
}
