% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{listDatasets}
\alias{listDatasets}
\title{List Datasets}
\usage{
listDatasets(limit = 50, offset = 0, sort = NULL)
}
\arguments{
\item{limit}{Maximum number of entries to return}

\item{offset}{Starting index into the result}

\item{sort}{Sort the results column, e.g. 'name'}
}
\value{
A \code{data.frame} containing the available datasets
}
\description{
List the datasets available to you in Domo.
}
\seealso{
\url{https://developer.domo.com/docs/dataset-api-reference/dataset}
}
