% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{queryDataset}
\alias{queryDataset}
\title{Query Dataset}
\usage{
queryDataset(dataset_id, sql = "SELECT * FROM table LIMIT 10")
}
\arguments{
\item{dataset_id}{Dataset ID to query}

\item{sql}{to query the dataset}
}
\value{
A \code{list} containing \code{data} and \code{metadata} from the queried dataset.
}
\description{
Query a dataset by dataset ID with SQL.
}
\examples{
SQL dialect is MySQL (i think)
SELECT * FROM table LIMIT 10
SELECT DISTINCT `Device Category`, `Web Property`, `Account` FROM table
SELECT SUM(Pageviews) pv, sum(Sessions) sessions FROM table GROUP BY Hostname
SELECT * FROM table LIMIT 10, 25"), # Limit to 10 offset by 25, nice!!!
SELECT * FROM table WHERE period = 201907 LIMIT 500000, 500000"
}
\seealso{
\url{https://developer.domo.com/docs/dataset-api-reference/dataset}
}
